Rename PotentialAttackException namespace and base class.
