+++ src/Exception/Attack/AttackException.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tuf\Exception\PotentialAttackException;
+namespace Tuf\Exception\Attack;
 
 use Tuf\Exception\TufException;
 
@@ -10,6 +10,6 @@ use Tuf\Exception\TufException;
  * Use this class for all failures related to verifying trust in the remote
  * repository metadata or a remote target.
  */
-abstract class PotentialAttackException extends TufException
+abstract class AttackException extends TufException
 {
 }
