+++ src/Exception/Attack/DenialOfServiceAttackException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Tuf\Exception\Attack;
+
+/**
+ * Indicates a potential denial of service attack.
+ */
+class DenialOfServiceAttackException extends AttackException
+{
+}
