+++ src/Exception/Attack/FreezeAttackException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Tuf\Exception\Attack;
+
+/**
+ * Indicates a potential freeze attack.
+ */
+class FreezeAttackException extends AttackException
+{
+}
