+++ src/Exception/Attack/RollbackAttackException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Tuf\Exception\Attack;
+
+/**
+ * Indicates a potential rollback attack.
+ */
+class RollbackAttackException extends AttackException
+{
+}
