+++ src/Exception/Attack/SignatureThresholdExpception.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Tuf\Exception\Attack;
+
+/**
+ * Indicates a resource does not reach the signature threshold.
+ */
+class SignatureThresholdExpception extends AttackException
+{
+}
