--- src/Exception/PotentialAttackException/DenialOfServiceAttackException.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Tuf\Exception\PotentialAttackException;
-
-/**
- * Indicates a potential denial of service attack.
- */
-class DenialOfServiceAttackException extends PotentialAttackException
-{
-}
