--- src/Exception/PotentialAttackException/FreezeAttackException.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Tuf\Exception\PotentialAttackException;
-
-/**
- * Indicates a potential freeze attack.
- */
-class FreezeAttackException extends PotentialAttackException
-{
-}
