--- src/Exception/PotentialAttackException/RollbackAttackException.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Tuf\Exception\PotentialAttackException;
-
-/**
- * Indicates a potential rollback attack.
- */
-class RollbackAttackException extends PotentialAttackException
-{
-}
