--- src/Exception/PotentialAttackException/SignatureThresholdExpception.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Tuf\Exception\PotentialAttackException;
-
-/**
- * Indicates a resource does not reach the signature threshold.
- */
-class SignatureThresholdExpception extends PotentialAttackException
-{
-}
