+++ src/Metadata/Verifier/UniversalVerifier.php
@@ -60,10 +60,10 @@ class UniversalVerifier
      * @param \Tuf\Metadata\MetadataBase $untrustedMetadata
      *   The untrusted metadata object.
      *
-     * @throws \Tuf\Exception\PotentialAttackException\FreezeAttackException
-     * @throws \Tuf\Exception\PotentialAttackException\RollbackAttackException
-     * @throws \Tuf\Exception\PotentialAttackException\InvalidHashException
-     * @throws \Tuf\Exception\PotentialAttackException\SignatureThresholdExpception
+     * @throws \Tuf\Exception\Attack\FreezeAttackException
+     * @throws \Tuf\Exception\Attack\RollbackAttackException
+     * @throws \Tuf\Exception\Attack\InvalidHashException
+     * @throws \Tuf\Exception\Attack\SignatureThresholdExpception
      */
     public function verify(string $role, MetadataBase $untrustedMetadata): void
     {
