+++ tests/Client/UpdaterTest.php
@@ -10,9 +10,9 @@ use Tuf\Client\Updater;
 use Tuf\Exception\DownloadSizeException;
 use Tuf\Exception\MetadataException;
 use Tuf\Exception\NotFoundException;
-use Tuf\Exception\PotentialAttackException\InvalidHashException;
-use Tuf\Exception\PotentialAttackException\RollbackAttackException;
-use Tuf\Exception\PotentialAttackException\SignatureThresholdExpception;
+use Tuf\Exception\Attack\InvalidHashException;
+use Tuf\Exception\Attack\RollbackAttackException;
+use Tuf\Exception\Attack\SignatureThresholdExpception;
 use Tuf\Exception\RepoFileNotFound;
 use Tuf\Exception\TufException;
 use Tuf\Metadata\RootMetadata;
