+++ tests/Unit/Client/VerifierTest.php
@@ -62,7 +62,7 @@ class VerifierTest extends TestCase
      */
     public function testCheckRollbackAttackAttack(): void
     {
-        $this->expectException('\Tuf\Exception\PotentialAttackException\RollbackAttackException');
+        $this->expectException('\Tuf\Exception\Attack\RollbackAttackException');
         $this->expectExceptionMessage('Remote any metadata version "$1" is less than previously seen any version "$2"');
 
         // The incoming version is lower than the local version, so this should
@@ -100,7 +100,7 @@ class VerifierTest extends TestCase
      */
     public function testCheckRollbackAttackAttackExpectedVersion(): void
     {
-        $this->expectException('\Tuf\Exception\PotentialAttackException\RollbackAttackException');
+        $this->expectException('\Tuf\Exception\Attack\RollbackAttackException');
         $this->expectExceptionMessage('Remote \'root\' metadata version "$2" does not the expected version "$3"');
 
         // The incoming version is lower than the local version, so this should
@@ -167,7 +167,7 @@ class VerifierTest extends TestCase
      */
     public function testCheckFreezeAttackAttack(): void
     {
-        $this->expectException('\Tuf\Exception\PotentialAttackException\FreezeAttackException');
+        $this->expectException('\Tuf\Exception\Attack\FreezeAttackException');
 
         $signedMetadata = $this->getMockBuilder(MetadataBase::class)->disableOriginalConstructor()->getMock();
         $signedMetadata->expects(self::any())->method('getType')->willReturn('any');
